Makes the add button for new APNs appear in the action bar.
The APN add button (a plus sign) now appears in the action bar if possible. The APN list interface is now more consistent with the WiFi list interface. Previously the user had to press the 'Menu' hardware button or the three dot button on the screen in order to reveal the actions.

Change-Id: Iae622054ec080ba4775c22223dd75a865c83fc27
Signed-off-by: Manuel R. Ciosici <manuelrciosici@gmail.com>