Changed the way that date format selector is shown (revision)
Changed the way that date format selector is shown excluding ambiguity when day and month have same value.
- i.e. 01/01/2012 (mm-DD-YYYY) and 01/01/2012 (DD-mm-YYYY)
Now it displays 31/12/2012 for DD-mm-YYYY and 12/31/2012 for mm-DD-YYYY.

This revision of previous code will also prevents on some timezones to display 1st january of next year instead 31th december of current year.

Change-Id: I9071341753e3da34da590bfef049ddbb2d635293