Properly update Vibrate when ringing preference
Previously when the ringing mode was changed, or onResume() was called,
mVibrateWhenRinging would get updated. However, only the vibrate mode was
being checked, and it was using deprecated APIs.

Now the setting is enabled always if the user has set the setting, or when
the phone is in vibrate mode.

Signed-off-by: Roman Birg <romanbirg@gmail.com>