Separate Change Encryption Password Entry (1 of 2)
Separates the link between changing the lockscreen password and changing the device encryption password.
This permits the use of a longer, more secure, password when deriving the filesystem encryption key,
yet doesn't affect ease-of-use when regularly unlocking the screen, as a short password or PIN
can be used. Patch requires framework modification as well as this, to function.

(ported to cm11 - cyanogen)

Change-Id: Ie2da961d87b5ca7286dd72562e7d9fbdf5bf954d

Conflicts:
	AndroidManifest.xml
	res/values/cm_strings.xml