QS: make dynamic tiles configurable (2/2)
instead of using a bunch of provider settings var
like it was on 4.3 and almost all other roms are doing
use one and split like we do for the normal tiles. This
is more consistent and saves us to observe a bunch of values

As well this is integrated into the one layout screen we have. So
all QS related is easy and logical usable for the user on one
screen

PatchSet: added icon - thanks to kroz

Change-Id: I32756584018585644a71bce52e48a978993d27ce

Conflicts:
	res/values/slim_strings.xml