Settings: handle stylus options only on full language settings
there are two states in language and input settings. On the one hand
the normal which is accessable from the settings and the one
if the user press the button for settings in the IME switcher popup.
This removes all default views and add dynamic ones. Due that Stylus
option do not belong to the dynamic view it will end up in a
crash (NPE see removeAllViews() before)

so check which status settings were called and do the work or just skip
it

Change-Id: Ic7f051d36bc88c3f23de5262e8413c79699afa1b

Conflicts:
	src/com/android/settings/inputmethod/InputMethodAndLanguageSettings.java