Settings: Add user protection and pie/navbar toggle to SlimActions
Add drawables for it and enable it for navbar, pie, hwkeys
and GlobalMenu and whatever other shorcutfeature want to have it
Keep the user save and do not allow to disable all navigations to keep
the device usable. User will be notified

In Settings user get the choice to enable the other navigation method. Of course
only on navbar devices due that hw key devices always have a working navigation

and a small duplicate cleanup

PS:
- added Jubas suggestions

Change-Id: I3d13a13bf54304b0207b83b54df5c37337bcdaf5

Conflicts:
	res/values/slim_strings.xml